+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release 2020.5 (?)
 
 ## Bug fixes
+- Fixed import loop when using `inmanta.execute.proxy` as entry point (#2341)
 - Clearing an environment with merged compile requests no longer fails (#2350)
 
 # Release 2020.4 (2020-09-08)